refactor(utilities): moving utilities package contents to contextually closer packages
