+++ go/data/utilities/strings.go
@@ -1,0 +1,16 @@
+package utilities
+
+import (
+	"strings"
+
+	"github.com/AssetMantle/schema/go/data/constants"
+)
+
+// TODO write testcase for empty and singular input
+func JoinListStrings(listStrings ...string) string {
+	return strings.Join(listStrings, constants.ListStringSeparator)
+}
+
+func SplitListString(listString string) []string {
+	return strings.Split(listString, constants.ListStringSeparator)
+}
