+++ go/utilities/codec.go
@@ -1,7 +1,4 @@
-// Copyright [2021] - [2022], AssetMantle Pte. Ltd. and the code contributors
-// SPDX-License-Identifier: Apache-2.0
-
-package codec
+package utilities
 
 import (
 	"reflect"
